old-tests/submission/Infraware/Forms/contents/Forms/textarea_wrap.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>textarea_wrap</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<form method="post"
		  enctype="application/x-www-form-urlencoded"
		  action=""
		  name="input_form">
	<p><textarea id='textarea_id'></textarea></p>
	</form>

	<script>
		
		var textarea = document.getElementById("textarea_id");
		textarea.wrap = 'Soft';
		var wrap1 = textarea.wrap;
		textarea.wrap = 'Hard';
		var wrap2 = textarea.wrap;

		if (typeof(textarea.wrap) == "string") {
			if (wrap1 == 'Soft' && wrap2 == "Hard") {
				test( function() {
					assert_true(true, "wrap attribute is correct.");
				});
			} else {
				test( function() {
					assert_unreached("wrap attribute is not correct.");
				});
			}
		} else {
			test( function() {
				assert_unreached("wrap attribute is not exist.");
			});
		}

	</script>

 </body>
</html>
